Making dataset double() by default

+++ utils/dataset.py
@@ -8,8 +8,8 @@ class Dataset:
     def __init__(self, X, y, verbose=True):
         self.device = torch.device("cuda" if torch.cuda.is_available()
                                    else "cpu")
-        self.X = X.to(self.device)
-        self.y = y.to(self.device)
+        self.X = X.to(self.device).double()
+        self.y = y.to(self.device).double()
 
         if verbose:
             print("Number of rows in X: {}".format(self.X.size(0)))
